services/analyzer/requirements.txt
celery[redis]==5.3.6
vaderSentiment==3.3.2
psycopg2-binary==2.9.9
pydantic==2.6.1
python-dotenv==1.0.1
